# saas/openapi/data-plane.yaml
openapi: 3.1.0
info:
  title: SaaS Data Plane API
  version: 0.0.3
  description: Per-deployment data plane API. One vhost per Deployment under *.saas.omarss.net.
servers:
  - url: https://{env}.{project}.saas.omarss.net
    variables:
      env:
        default: dev
        description: Environment slug (dev, staging, prod).
      project:
        default: example
        description: Project slug for the Deployment.
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List tenants visible to the caller's Deployment.
      description: |
        Returns the tenants the caller is authorised to see. In Phase 2 the
        Data Plane caller's token resolves to exactly one tenant, so the list
        contains one element. Operator impersonation across tenants lands in
        Phase 13.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "410":
          $ref: "#/components/responses/CursorGone"
        "429":
          $ref: "#/components/responses/RateLimited"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a tenant. Auto-creates a default Organization.
      description: |
        Idempotent. Replays with the same Idempotency-Key + body return the
        cached 201 response; with the same key + a different body return 422
        idempotency-key-conflict (see AGENTS.md section 5.2).
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Created
          headers:
            ETag:
              schema:
                type: string
              description: 'Weak ETag, format W/"v<sequence>".'
            Location:
              schema:
                type: string
              description: URI of the created tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/IdempotencyInFlight"
        "422":
          $ref: "#/components/responses/IdempotencyConflictOrValidation"
...
  /v1/tenants/{tenant_id}/audit-events:
    parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          type: string
          pattern: "^tenant_[0-9A-HJKMNP-TV-Z]{26}$"
    get:
      operationId: listAuditEvents
      tags: [audit]
      summary: List audit events for a tenant.
      description: |
        Returns audit events ordered by occurred_at DESC. MVP gates this
        endpoint with AssertTenant only; the audit.read permission lands
        with the v1 RBAC hardening pass (ADR 012).
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: resource_type
          schema: { type: string }
        - in: query
          name: resource_id
          schema: { type: string }
        - in: query
          name: actor_id
          schema: { type: string }
        - in: query
          name: occurred_after
          schema: { type: string, format: date-time }
        - in: query
          name: occurred_before
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/audit-events/{audit_event_id}:
    parameters:
      - in: path
        name: audit_event_id
        required: true
        schema:
          type: string
          pattern: "^audit_[0-9A-HJKMNP-TV-Z]{26}$"
    get:
      operationId: getAuditEvent
      tags: [audit]
      summary: Fetch a single audit event by id.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/audit-events/export:
    post:
      operationId: exportAuditEvents
      tags: [audit]
      summary: Export filtered audit events as JSON or CSV.
      description: |
        Synchronous export bounded to ExportSyncBudget (1 MiB). Larger
        exports return 413; the async export path (background worker
        + signed URL) lands with the v1 Files module.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportAuditEventsRequest"
      responses:
        "200":
          description: Inline export.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventListResponse"
            text/csv:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          description: Export exceeds sync size budget.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
components:
  schemas:
...
    Tenant:
      type: object
      required:
        - id
        - object
        - slug
        - name
        - status
        - created_at
        - updated_at
        - etag
      properties:
        id:
          type: string
          pattern: "^tenant_[0-9A-HJKMNP-TV-Z]{26}$"
        object:
          type: string
          enum: [tenant]
        slug:
          type: string
          minLength: 2
          maxLength: 32
          pattern: "^[a-z][a-z0-9-]*[a-z0-9]$"
        name:
          type: string
          minLength: 1
          maxLength: 120
        status:
          type: string
          enum: [active, suspended, deleted]
        default_organization_id:
          type: string
          nullable: true
          pattern: "^org_[0-9A-HJKMNP-TV-Z]{26}$"
        metadata:
          $ref: "#/components/schemas/Metadata"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
        etag:
          type: string
          description: 'Weak ETag, format W/"v<sequence>".'
    TenantResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Tenant"
    TenantListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Tenant"
        pagination:
          $ref: "#/components/schemas/Pagination"
    CreateTenantRequest:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
          minLength: 2
          maxLength: 32
          pattern: "^[a-z][a-z0-9-]*[a-z0-9]$"
        name:
          type: string
          minLength: 1
          maxLength: 120
        metadata:
          $ref: "#/components/schemas/Metadata"
...
    Metadata:
      type: object
      maxProperties: 16
      additionalProperties:
        type: string
        maxLength: 256
...
    AuditEvent:
      type: object
      required:
        - id
        - tenant_id
        - actor_type
        - actor_id
        - action
        - resource_type
        - occurred_at
        - prev_hash
        - row_hash
        - chain_sequence
      properties:
        id:
          type: string
          pattern: "^audit_[0-9A-HJKMNP-TV-Z]{26}$"
        tenant_id:
          type: string
        actor_type:
          type: string
          enum: [user, api_key, operator, operator_impersonation, system]
        actor_id:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
          nullable: true
        occurred_at:
          type: string
          format: date-time
        ip_address:
          type: string
          nullable: true
          x-oapi-codegen-extra-tags:
            pii: "true"
        user_agent:
          type: string
          nullable: true
          x-oapi-codegen-extra-tags:
            pii: "true"
        request_id:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        prev_hash:
          type: string
          description: Lowercase hex SHA-256 of the previous row, or the per-tenant Genesis seed for chain_sequence=1.
        row_hash:
          type: string
          description: Lowercase hex SHA-256 of (prev_hash || JCS(canonical row body)).
        chain_sequence:
          type: integer
          format: int64
          description: Per-tenant monotonic sequence number starting at 1.
    AuditEventResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AuditEvent"
    AuditEventListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
        pagination:
          $ref: "#/components/schemas/Pagination"
    ExportAuditEventsRequest:
      type: object
      properties:
        tenant_id:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        actor_id:
          type: string
        occurred_after:
          type: string
          format: date-time
        occurred_before:
          type: string
          format: date-time
        format:
          type: string
          enum: [json, csv]
          default: json
        limit:
          type: integer
          minimum: 1
          maximum: 10000
    Pagination:
      type: object
      required: [has_more]
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: string
  parameters:
    Limit:
      in: query
      name: limit
      required: false
      description: Max items to return (default 25, max 200).
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    Cursor:
      in: query
      name: cursor
      required: false
      description: Opaque pagination cursor; obtained from a previous response.
      schema:
        type: string
    Sort:
      in: query
      name: sort
      required: false
      description: Sort token. Default "-created_at".
      schema:
        type: string
        default: "-created_at"
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: 24-hour idempotency key (idem_<ulid>).
      schema:
        type: string
        pattern: "^idem_[0-9A-HJKMNP-TV-Z]{26}$"
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token / API key.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Caller lacks permission for this resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    IdempotencyInFlight:
      description: Concurrent request with the same Idempotency-Key is still processing.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    IdempotencyConflictOrValidation:
      description: Idempotency-Key reused with a different body, OR validation failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    CursorGone:
      description: Cursor schema version is no longer supported.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: Rate limit exceeded.
      headers:
        RateLimit-Limit:
          schema:
            type: integer
        RateLimit-Remaining:
          schema:
            type: integer
        RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API key